% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Baseline_Correction.R
\name{Baseline_Correction}
\alias{Baseline_Correction}
\title{Baseline_Correction}
\usage{
Baseline_Correction(lefichier = NULL)
}
\arguments{
\item{lefichier}{a filename to work on.}
}
\value{
Baseline corrected spectra and baseline spectra are stored in files
   with the same name as the input file but with "Corrected" or "Baseline"
   keyword added.
}
\description{
A simple GUI to remove a baseline from spectra in a file.
}
\details{
The file format must conform to the one required by InSpectoR.
See \link{InSpectoR} for details. The method to calculate the
baseline is from the \link{baseline} package, namely the baseline
function with \emph{method='modpolyfit'}. A GUI helps the user
to select parameters to perform the fit.
}
\examples{
dfile <- system.file("raman","Raman_I.txt",package="inspectrar")
Baseline_Correction(dfile)
}
