% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP_Scatter.R
\name{BP_Scatter}
\alias{BP_Scatter}
\title{BP_Scatter}
\usage{
BP_Scatter(
  ax1,
  ax2,
  labs = list("X", "Y"),
  titre = "",
  xlimits = NULL,
  ylimits = NULL,
  colorby = rep(2, length(ax1)),
  pt.label = NULL
)
}
\arguments{
\item{ax1}{vector of values on the horizontal axis}

\item{ax2}{vector of values on the vertical axis}

\item{labs}{2 element list giving x and y axis labels respectively}

\item{titre}{main figure title}

\item{xlimits}{minimum and maximum values for the horizontal axis}

\item{ylimits}{minimum and maximum values for the vertical axis}

\item{colorby}{class for color coding of points. Default is all blue. 
When default, no legend is plotted.}

\item{pt.label}{label for points. NA is used to skip a labels.}
}
\value{
A list with the following elements:
 \itemize{
         \item Xsc,Ysc  : x and y values of all scores plotted.
             \item  labs : 2 element list giving x and y axis labels respectively
        \item  titre     : main figure title
    \item  X_Lim, Y_Lim  : xlim and ylim of current plot.
         \item  colorby  : class for color coding of points.
         \item  pt.label : label for points. NA is used to skip a labels.
         }
}
\description{
Produces a scatter plot and return parameters for easy replotting when zooming for example.
}
\examples{
BP_Scatter(ax1=c(rnorm(20,10,2),rnorm(20,20,3)), ax2=rnorm(40,5,3),
                labs=list("X-axis","Y-axis"),
                titre = "A scatter plot",
                xlimits=c(5,30), ylimits=c(2,8),
                colorby=as.factor(c(rep("Cl1",20),rep("Cl2",20))))
                
}
