% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP_Spectra_matplot.R
\name{BP_Spectra_matplot}
\alias{BP_Spectra_matplot}
\title{BP_Spectra_matplot}
\usage{
BP_Spectra_matplot(x, yID, col_code, subs, xlimits, ylimits, preT = FALSE)
}
\arguments{
\item{x}{list of matrices containing spectra. Each matrix is a different type
of spectra.}

\item{yID}{name of ASCII data files containing spectral matrices. A different
line type will be use to identify data source in plots.}

\item{col_code}{vector of color IDs. Usually, this will be a column (a
factor) in the Ys_df of the data set.}

\item{subs}{vector of row number in Y file of selected samples.}

\item{xlimits}{vector of xmin and xmax}

\item{ylimits}{vector of ymin and ymax}

\item{preT}{TRUE for pretreated data. FALSE otherwise.}
}
\value{
A plot on the active graphic device
flags    : list(xdefault, ydefault) indicating caller that x and y axis limits were set to default (TRUE).
}
\description{
BP_matplot is matplot with parameters tailored for use in InSpectoR. Called
by InSpectoR. Remember that data for plotting x-axis (e.g. wavelengths) are
in the first row of the matrices
}
