% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToClipboard.R
\name{copyToClipboard}
\alias{copyToClipboard}
\title{copyToClipboard}
\usage{
copyToClipboard(g)
}
\arguments{
\item{g}{the ggraphics widget}
}
\value{
The content of the ggraphics widget on the clipboard, ready to paste
}
\description{
Copies then content of a ggraphics widget to the clipboard
}
\examples{
 w=gWidgets2::gwindow(visible=FALSE)
 gg=gWidgets2::ggraphics(cont=w)
 gWidgets2::visible(w) <- TRUE
 graphics::hist(rnorm(2000))
 copyToClipboard(gg)
 # Paste somewhere!
 
}
