% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_limits.R
\name{get_graph_limits}
\alias{get_graph_limits}
\title{get_graph_limits}
\usage{
get_graph_limits(x)
}
\arguments{
\item{x}{a list of spectral data files. Usually XData or XData_p}
}
\value{
A list with two elements
\itemize{
\item xlimits: a vector c(xmin,xmax)
\item ylimits: a vector c(ymin,ymax)}
}
\description{
A helper function for InSpectoR. To get plotting limits from a list of spectral 
data files. Not exported.
}
\examples{
dfile <- system.file("foodstuff_powder","Y_foodstuff.txt",package="inspectrar")
InSpectoR(dfile,0,1600,1024)
#**************************************************
#GUI may wait for your answer for creating a backup!
#**************************************************
leslimites <- get_graph_limits(XData_p)
#******************************************
#Don't forget that an active GUI is running.
#******************************************

}
