% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InSpectoR.R
\name{InSpectoR}
\alias{InSpectoR}
\title{InSpectoR}
\usage{
InSpectoR(yfile = NULL, parcomp = TRUE, MainWidth = 1200,
  MainHeight = 800)
}
\arguments{
\item{yfile}{to load data associated with this Y data file.
Default is NULL: no data loaded on startup.}

\item{parcomp}{to set up for parallel computation with caret package
(default to TRUE)}

\item{MainWidth}{width of GUI in pixels (default to 1200)}

\item{MainHeight}{height of GUI in pixels (default to 800)}
}
\value{
NOTHING
}
\description{
InSpectoR is a tool for exploring spectroscopic data sets and to develop
models. It was developped as a companion to a custom spectrometric data
acquisition system for gathering induced fluorescence with up to 6 
excitation wavelengths, Raman spectra excited at ~785 nm and transmittance
data or reflectance data. It is general enough to accomodate any spectroscopic
data sets provided that the format of the required files conformed to the
description below. The program is structured around a top level window with tabs giving
access to other interfaces dedicated to specialized tasks such as selecting
and viewing raw data, performing data normalization, doing some
pretreament on the data, performing principal component analysis,
developping PLS models... These are referred to as "task interfaces".
}
\details{
The data sets are made of a number of files defined as follows:
\itemize{
  \item Y_somename.txt  : Y data file.
    \itemize{
     \item This is a tab delimited text file with a header line and decimal symbol is ".". 
    \item After header, each following lines correspond to samples.
    \item Columns contain user defined data such as a sample
    identification number, some data or factor associated
    with the sample (e.g. color, pH, concentration of
                 something...). 
    \item The filename MUST start with "Y_". 
    It is followed by "somename" which is a user defined
    name (anything will do) and followed by the ".txt"
    extension. 
    \item The GUI offers access to a support function
    that can merge Y data in an other file. This is useful when
    spectral data and accompanying data are acquired in separate
    processes that might result in data stored in different files.
    The only requirement for successfull merging is that both
    files share a common column.
    }
 \item *_somename*.txt : Spectral data files. Tab delimiter, header and decimal=".".
    \itemize{
    \item These are files containing spectroscopic data.
    \item The first line contains wavelengths or wavenumbers
    associated with the spectra. 
    \item The first column is a sample ID (same as first column of
                    associated Y file). 
    \item One sample per line.
    \item The first cell [1,1] is not used (could be empty).
    \item Name: The first star to the left is a placeholder for
    describing the data. For example, EX350 could be used
    to indicate fluorescence spectra obtained under
    excitation at 350 nm. These data type are used throughout
    the program as labels for options, buttons... Should not
    start with a number or special character.
    It is up to the user to define
    and recognize the names. The * before the extension
    could be anything. The key is that "somename" matches
    the "somename" of the associated Y_somename.txt file.
    This is how Y data are connected to spectral data.
    }
 }
 
    It is mandatory to group Ys and spectral data files
    in the same directory. There could be as many spectral
    data files as the user wants.
}
\examples{
dfile <- system.file("foodstuff_powder","Y_foodstuff.txt",package="inspectrar")
InSpectoR(dfile,0,1600,1024)
}
