% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match_Dataset_Multiple.R
\name{Match_Dataset_Multiple}
\alias{Match_Dataset_Multiple}
\title{Match_Dataset_Multiple}
\usage{
Match_Dataset_Multiple(...)
}
\value{
Nothing. Creates a new set of data files (Y and X files) of the
 merged data
}
\description{
Undocumented function to merge data in the special case where several 
instruments are used to acquire data but the data types are not the
same for each of the instrument.
}
\examples{
Match_Dataset_Multiple()

}
