% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_4_pls.R
\name{R2.mvr}
\alias{R2.mvr}
\title{R2_4_pls}
\usage{
\method{R2}{mvr}(object, estimate, newdata, ncomp = 1:object$ncomp,
  comps, intercept = cumulative, se = FALSE, ...)
}
\arguments{
\item{object}{}

\item{estimate}{}

\item{newdata}{}

\item{ncomp}{}

\item{comps}{}

\item{intercept}{}

\item{se}{}

\item{...}{}
}
\value{

}
\description{
NORMALLY, THIS FUNCTION IS PROVIDED BY THE PLS PACKAGE.
HAD SOME PROBLEM WITH A GIVEN (??) VERSION OF THE PLS PACKAGE.
THIS CODE WAS COPIED FROM THE PLS PACKAGE VERSION 2.7-1.
Helper function to get R^2 values for pls fits.
Depends on pls package.
}
