% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ref_2_Lab.R
\name{Ref_2_Lab}
\alias{Ref_2_Lab}
\title{Ref_2_Lab}
\usage{
Ref_2_Lab(wS, S)
}
\arguments{
\item{wS}{wavelenth vector}

\item{S}{matrix of reflectance spectra. One spectrum per line. Number of columns should match length(wS)}
}
\value{
A matrix of dimension (nrow(S),3) where the columns are L, a and b value, one row per sample
}
\description{
Function to convert reflectance spectra to CIELab values
}
