% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remove_line_TextFile.R
\name{Remove_line_TextFile}
\alias{Remove_line_TextFile}
\title{Remove_line_TextFile}
\usage{
Remove_line_TextFile(oldfile, line_2_del, newfile = oldfile)
}
\arguments{
\item{oldfile}{text file to modify by removing lines}

\item{line_2_del}{a vector of line numbers to remove}

\item{newfile}{text file. This is \strong{oldfile} after line removal}
}
\value{
NOTHING
}
\description{
Function to remove lines from a text file.
}
