% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trouve_Err_EchID.R
\name{Trouve_ErrEchID}
\alias{Trouve_ErrEchID}
\title{Trouve_ErrEchID}
\usage{
Trouve_ErrEchID(fromConsole = T)
}
\arguments{
\item{fromConsole}{a boolean. TRUE for running from console. FALSE for running from InSpectoR}
}
\value{
\itemize{
     \item a character string ("Compatible files") when files are compatible 
     \item when files are not compatible, a list having three elements:
     \itemize{
         \item a data frame with three columns: 1-line where non matchings appear, 
                                                2-Y file value in column 1 for line in item 1 above,
                                                3-X file value in column 1 for line in item 1 above.
         \item the x full file name
         \item the y full file name
     }
   }
}
\description{
Function to identify non matching items in the first columns of Y and X files.
}
\examples{
Trouve_ErrEchID()

}
